Cast Slider exepected value to int (#13840)

+++ src/Platform.Handlers/tests/Xamarin.Platform.Handlers.DeviceTests/Handlers/Slider/SliderHandlerTests.Android.cs
@@ -70,7 +70,7 @@ namespace Xamarin.Platform.Handlers.DeviceTests
 				Value = xplatValue
 			};
 
-			double expectedValue = SliderExtensions.NativeMaxValue / 2;
+			int expectedValue = (int)(SliderExtensions.NativeMaxValue / 2);
 
 			var values = await GetValueAsync(slider, (handler) =>
 			{
